# src/openapi/schema.yaml
openapi: 3.0.0
info:
  title: Recipes AI API
  version: 1.0.0
  description: API for Recipes AI service
paths:
  /health:
    get:
      summary: Health check endpoint
      description: Returns the status of the service
      operationId: getHealth
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [ok]
                  timestamp:
                    type: string
                    format: date-time
                required:
                  - status
                  - timestamp
  /ingredients:
    post:
      summary: Add ingredients for a recipe
      description: Adds ingredients for a recipe to the vector store
      operationId: addIngredients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recipe_id:
                  type: string
                  description: ID of the recipe
                ingredients:
                  type: string
                  description: Ingredients list as text
              required:
                - recipe_id
                - ingredients
      responses:
        '201':
          description: Ingredients added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  message:
                    type: string
                  id:
                    type: string
                required:
                  - success
                  - message
                  - id
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [false]
                  error:
                    type: string
                required:
                  - success
                  - error
    delete:
      summary: Delete ingredients by recipe ID
      description: Deletes all ingredients associated with a recipe ID. Use recipe_id='*' to delete all ingredients from the collection.
      operationId: deleteIngredients
      parameters:
        - name: recipe_id
          in: query
          required: true
          schema:
            type: string
          description: ID of the recipe to delete ingredients for. Use '*' to delete all ingredients.
      responses:
        '200':
          description: Ingredients deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  message:
                    type: string
                  deletedCount:
                    type: integer
                    description: Number of ingredients deleted
                required:
                  - success
                  - message
                  - deletedCount
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [false]
                  error:
                    type: string
                required:
                  - success
                  - error
  /ingredients/search:
    post:
      summary: Search for recipes by ingredients
      description: Searches for recipes with similar ingredients
      operationId: searchIngredients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ingredients:
                  type: string
                  description: Ingredients to search for
                limit:
                  type: integer
                  description: Maximum number of results to return
                  default: 3
                  minimum: 1
                  maximum: 20
              required:
                - ingredients
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        recipe_id:
                          type: string
                        ingredients:
                          type: string
                      required:
                        - recipe_id
                        - ingredients
                required:
                  - success
                  - results
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [false]
                  error:
                    type: string
                required:
                  - success
                  - error
  /ingredients/batch:
    post:
      summary: Batch upload ingredients for multiple recipes
      description: Adds ingredients for multiple recipes to the vector store in a single operation
      operationId: addIngredientsBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                records:
                  type: array
                  description: Array of ingredient records to add
                  items:
                    type: object
                    properties:
                      recipe_id:
                        type: string
                        description: ID of the recipe
                      ingredients:
                        type: string
                        description: Ingredients list as text
                    required:
                      - recipe_id
                      - ingredients
              required:
                - records
      responses:
        '201':
          description: Ingredients batch added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  message:
                    type: string
                  ids:
                    type: array
                    items:
                      type: string
                required:
                  - success
                  - message
                  - ids
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [false]
                  error:
                    type: string
                required:
                  - success
                  - error 